backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.0.5
python-dotenv==1.0.0
bcrypt==4.0.1
PyJWT==2.12.1
Werkzeug==2.3.7
PyMySQL==1.1.0
mysql-connector-python==8.0.33
cryptography==41.0.4
